+++ source/js/svg4everybody-init.js
@@ -1,0 +1,1 @@
+svg4everybody();
