+++ source/sass/blocks/catalog.scss
@@ -55,6 +55,6 @@
   }
 
   .catalog__zigzag-bg {
-    top: -145px;
+    top: -147px;
   }
 }
