+++ source/sass/blocks/category-list.scss
@@ -44,6 +44,7 @@
 
 .category-list__item {
   width: 130px;
+  margin: 0;
 }
 
 .category-list__item::after {
