+++ source/sass/blocks/contacts.scss
@@ -95,6 +95,8 @@
 .contacts__map {
   display: flex;
   width: 100%;
+
+  border: none;
 }
 
 .contacts__write-button {
