+++ source/sass/blocks/main-nav.scss
@@ -1,3 +1,7 @@
+.main-nav {
+  width: 100%;
+}
+
 .main-nav__list {
   margin: 0;
   padding: 0;
@@ -76,27 +80,44 @@
 }
 
 @media (max-width: $mobile-width-only) {
+  .main-nav--opened .main-nav__wrapper {
+    display: block;
+  }
+  
   .main-nav--nojs .main-nav__button {
     display: none;
+  }
+
+  .main-nav--closed .main-nav__wrapper {
+    display: none;
+  }
+
+  .main-nav--nojs .main-nav__wrapper {
+    display: block;
   }
 }
 
 @media (min-width: $tablet-width) {
   .main-nav {
-    width: auto;
+    width: 100%;
+    height: 0;
+    padding: 0 $tablet-padding;
     margin: 0;
   }
 
   .main-nav__wrapper {
+    position: relative;
+
     display: flex;
     justify-content: space-between;
+    width: $tablet-width - $tablet-padding * 2;
+    margin: 0 auto;
 
     height: 88px;
   }
 
   .main-nav__wrapper--specials {
-    position: absolute;
-    right: 0;
+    position: relative;
   }
 
   .main-nav__button {
@@ -104,20 +125,35 @@
   }
 
   .main-nav__label {
-    display: flex;
+    position: absolute;
+    top: -69px;
+    right: 0;
     z-index: 3;
-
-    padding-top: 18px;
+    display: flex;
+   
+    margin-left: auto;
   }
 }
 
 @media (min-width: $desktop-width) {
   .main-nav {
     width: $desktop-width - $desktop-padding * 2;
+    height: 88px;
     margin: 0 auto;
+    padding: 0;
+  }
+
+  .main-nav__wrapper {
+    width: 1047px;
+    margin: 0;
   }
 
   .main-nav__wrapper--specials {
-    width: $desktop-width - $desktop-padding * 2;
+    width: 1047px;
+  }
+
+  .main-nav__label {
+    position: unset;
+    padding-top: 18px;
   }
 }
