+++ source/sass/blocks/page-header.scss
@@ -75,6 +75,7 @@
   .page-header__logo {
     position: absolute;
     left: 448px;
+    z-index: 3;
 
     width: auto;
     height: 90px;
