+++ source/sass/blocks/page-main.scss
@@ -1,7 +1,7 @@
 .page-main__header {
   width: 250px;
   padding-top: 70px;
-  padding-bottom: 104px;
+  padding-bottom: 122px;
   margin-top: 0;
   margin-right: auto;
   margin-left: auto;
@@ -27,11 +27,13 @@
 }
 
 .page-main__container-inner {
+  background-color: $color-white;
   background-image: url("../img/hero-mobile@1x.jpg");
   background-repeat: no-repeat;
   background-position: center -18px;
 
   @media (min-resolution: $retina-dppx) {
+    background-color: $color-white;
     background-image: url("../img/hero-mobile@2x.jpg");
     background-position: center -18px;
     background-size: 575px auto;
@@ -62,7 +64,7 @@
   .page-main__header {
     width: 500px;
     padding-top: 153px;
-    padding-bottom: 354px;
+    padding-bottom: 373px;
 
     font-size: 43px;
     line-height: 50px;
@@ -104,7 +106,7 @@
 
   .page-main__header {
     width: 680px;
-    padding-bottom: 340px;
+    padding-bottom: 359px;
 
     font-size: 53px;
     line-height: 57px;
