+++ source/sass/blocks/user-list.scss
@@ -1,6 +1,13 @@
 .user-list__link {
   position: relative;
 
+  display: flex;
+  width: 260px;
+  margin: 0 auto;
+  padding-top: 24px;
+  padding-left: 45px;
+  padding-bottom: 24px;
+
   font-size: 17px;
   font-weight: 700;
   line-height: 24px;
@@ -11,8 +18,8 @@
 
 .user-list__icon {
   position: absolute;
-  top: 5px;
-  left: -45px;
+  top: 28px;
+  left: -0px;
 
   width: 21px;
   height: 18px;
@@ -40,51 +47,41 @@
   opacity: 0.3;
 }
 
-.user-list__item-wrap {
-  border-top: 1px solid $color-lightergrey;
-}
-
 .user-list__item {
-  width: 260px;
-  margin: 0 auto;
-  padding-top: 24px;
-  padding-left: 45px;
-  padding-bottom: 24px;
-}
-
-@media (max-width: $mobile-width-only) {
-  .main-nav--closed .user-list__item-wrap {
-    display: none;
-  }
+  border-top: 1px solid $color-lightergrey;
 }
 
 @media (min-width: $tablet-width) {
   .user-list {
+    position: absolute;
+    right: 0;
+    top: -88px;
+    z-index: 3;
+
     display: flex;
-    justify-content: space-between;
   }
 
   .user-list__item {
     width: 264px;
-    padding-top: 33px;
+    padding-top: 31px;
     padding-left: 0;
-    padding-bottom: 31px;
-  }
-
-  .user-list__item-wrap {
+    padding-bottom: 33px;
+  
     border: none;
   }
 
-  .user-list__item-wrap--search {
-    padding-top: 33px;
-
-    border-left: 1px solid $color-lightergrey;
-    border-right: 1px solid $color-lightergrey;
+  .user-list__link {
+    width: auto;
+    margin: 0;
+    padding: 0;
   }
 
   .user-list__item--search {
     width: 82px;
-    padding-top: 0;
+    padding-top: 7px;
+
+    border-left: 1px solid $color-lightergrey;
+    border-right: 1px solid $color-lightergrey;
   }
 
   .user-list__icon--search {
@@ -107,6 +104,10 @@
 }
 
 @media (min-width: $desktop-width) {
+  .user-list {
+    position: unset;
+  }
+
   .user-list__item--cart {
     width: 267px;
   }
